Series boolean 비교 및 FutureWarning 완전 수정
문제:
1. Series is ambiguous 오류
   - Series 직접 비교 불가
2. FutureWarning: Calling float on Series
   - float(Series) deprecated

해결:
1. Series  float 변환
   - current_price, ma, deviation
   - boolean 비교 가능

2. .item() 메서드 사용
   - Series.item()  scalar
   - FutureWarning 제거

결과:
- 미국 시장 지표 정상 작동
- 모든 Warning 제거
- 완전 작동

+++ core/strategy/us_market_monitor.py
@@ -87,14 +87,14 @@ class USMarketMonitor:
             
             # 컬럼명 확인 (close 또는 Close)
             close_col = 'Close' if 'Close' in data.columns else 'close'
-            current_price = data[close_col].iloc[-1]
+            current_price = float(data[close_col].iloc[-1])
             
             # 이동평균 지표
             if 'period' in indicator_config:
                 period = indicator_config['period']
-                ma = data[close_col].rolling(period).mean().iloc[-1]
-                deviation = (current_price - ma) / ma
-                threshold = indicator_config.get('threshold', 0.02)
+                ma = float(data[close_col].rolling(period).mean().iloc[-1])
+                deviation = float((current_price - ma) / ma)
+                threshold = float(indicator_config.get('threshold', 0.02))
                 
                 # 신호 판단
                 if deviation > threshold:
